Updated/working code for the Register POST request
This update fixes the register method which is charge of make a POST
request to the UserReq.php server. It can now make a new user perfectly
:) - The only issue is with the PHP script, it still passes back 0,
even when the request has worked. So to summarise, the request works
and the new user account is created on the MySQL database, but the PHP
script still returns a 0 when it should be returning a 1.