Updates to Login, Register, Keychain & ReportBug
This commit adds support for Keychain access to securely store the
username and password. It adds a new Login/Register methods. They work
but are in need of some more work as the PHP for Login and register
don’t seem to work yet. This commit also contains further progress on
the Report Bug section of the app.